[LLVM][TableGen] Add JSON magic directive to Jupyter kernel
TODO: should this be some global setting? %nojson?

This allows you to see the records as pretty printed JSON,
which can give you a better idea of the structure.

This is equivalent to `%args --dump-json` except that
the kernel will format it before returning the result.

(and if you did both, it would still work)

Without formatting, the ouput is all one line.

An example is added to the end of the notebook
and is shown in the Markdown export.

Depends on D132646

Differential Revision: https://reviews.llvm.org/D132647