WIP: run python on json output
TODO: fix the merge conflict

This adds %py to run python on the json of the cached source.

Using eval means only one line, exec doesn't want to play well but that's
what we'd need if we're gonna do whole cells for this.

Either way security is a concern.

%py is going to have to imply the json option but that's fine I guess.

Sure would be nice if we could roundtrip json back into tablegen
and keep the cache that way.

Versus something like jq, I think the benfit is being able to write more complicated
things in standard syntax.

The drawback is the one liners get really weird, so we gotta fix that part.