[lldb][AArch64] Add register field information for SME's SVCR register
This register is a pseudo register but mirrors the architectural
register's contents. See:
https://developer.arm.com/documentation/ddi0616/latest/

For the full details. Example output:
```
(lldb) register read svcr
    svcr = 0x0000000000000002
         = (ZA = 1, SM = 0)
```

As this register is at this time only present on Linux, I have
added the field information in generic arm64 code.

For testing I have added check for the fields in
TestZARegisterSaveRestore.py. Though this test is a general SME
test, I don't see the point in the alternative, which is duplicating
it just to check fields are shown correctly.