[GitHub] Add greeting comment to opened PRs from new contributors
This includes some commonly needed information like how to add
reviewers.

This is implemented as a job before the labeler, so that on a new
PR the comment is added before there are any subscribers and only
the author gets a nofitication.

The labeler job depends on the greeter having run or having been skipped.
So if the PR wasn't just opened, or it's from a regular contributor,
the labeling still happens.

But we can be sure that when a greeting comment is left, it's the
very first thing we do.