[lldb] Parse and display register field enums
This teaches lldb to parse the XML sent by lldb-server,
and make use of the information in `register read` and
`register info`. The target XML parser will drop any
invalid enum or evalue. If we find multiple evalue
for the same value, we'll use the last one we find.

The order of evalues from the XML is preserved as
there may be good reason they are not in numerical
order.

`register info` will show enum fields where it finds them
and when printing we configure them to always show value and name
so you can compare against program source code easily.