Reland "[GitHub][workflows] Ask reviewers to merge PRs when author cannot (#81142)"
This reverts commit 124cd11d7f9d875b7d6dd441dc758ed9341ccec2.

The job originally failed because any workflow run on a PR runs
in a context that cannot write to the PR itself (otherwise people
could damage the repo using a workflow in a PR).

https://github.com/llvm/llvm-project/pull/80495 recently added
a job that is purely for commenting on issues, to solve a similair
problem. This limits the damage the PR can do to adding a spam comment.