[lldb] Show register descriptions in "register info"
If the register has a description it will be printed after the details,
before the field table.

Instead of intending this like the inital details I'm just printing it.
The terminal will wrap it automatically. Which I think makes more sense
for something which will be 1-2 sentences.

At this point, no register will have a description. Support for reading
them from XML is coming in a future change.