[llvm][llvm-readobj] Add NT_ARM_GCS Linux core note type
The ARM Guarded Control Stack extension (GCS) is similar to existing
shadow stack extensions for other architectures.

The core note will include which features of GCS are enabled,
which have been locked in their current state, and the stack
pointer of the shadow stack.

Note that 0x40e is NT_ARM_FPMR, which I will add in a future PR.