[lldb][Linux] Add Control Protection Fault signal
This will be sent by Arm's Guarded Control Stack extension
when an invalid return is executed.

The signal does have an address we could show, but it's the
PC at which the fault occured. The debugger has plenty of ways
to show you that already, so I've left it out.

```
(lldb) c
Process 460 resuming
Process 460 stopped
* thread #1, name = 'test', stop reason = signal SIGSEGV: control protection fault
    frame #0: 0x0000000000400784 test`main at main.c:57:1
   54  	  afunc();
   55  	  printf("return from main\n");
   56  	  return 0;
-> 57  	}
(lldb) dis
<...>
->  0x400784 <+100>: ret
```

The new test case generates the signal by corrupting the link
register then attempting to return. This will work whether we
manually enable GCS or the C library does it for us.

(in the former case you could just return from main and it would
fault)