[lldb] Introduce RegisterType base class for all register information classes
This is refactoring to prepare for https://github.com/llvm/llvm-project/issues/87471
where I will be adding unions.

https://sourceware.org/gdb/current/onlinedocs/gdb.html/Target-Description-Format.html

A union is like a C union and references other types defined in the XML. Just like
a set of register flags might reference an enum for one of those flags.

By introducing this base class I'm making the treatment of all these different
things generic. So that when encoding any of them to XML we:
* Emit the dependencies (and their dependencies and so on)
* Emit the element itself

This strategy will also be used later in RegisterTypeBuilderClang to generate
AST to represent these types.

As GDB decided to include size in enums, whenever we emit something it
will get a "user" pointer. This allows an enum to read the size of the
register it's being attached to. No other type class requires this.

LLDB does not care about this enum size attribute, but GDB does so we emit
it for compatibility.

We don't expect anything other than a register to reference an enum.
In theory we could be wrong, so I've defaulted to not emitting a size
if that happens.