[GitHub] Workflows to label PRs from authors without commit access...
...and then sweep open PRs to find approved PRs from authors
without commit access, and encourage their approvers to merge
on the author's behalf.

Ideally this would be done by one workflow that reacts to the
review approved event. However due to the source of that event,
workflows cannot modify the target repository in response. See:
https://github.com/orgs/community/discussions/26651
https://github.com/orgs/community/discussions/55940

So it was suggested to add a label on PR opening, which we
are allowed to do, then have a job on a timer convert
that into a comment if the PR was approved.

This means we will have some delay between approval and
asking for merge, but it's better than the days it can
take if neither side knows to ask.

The comment first tells the author to make final changes
to their PR. I have used the text from https://llvm.org/docs/Contributing.html#how-to-submit-a-patch
for this. We could add a specific link for that, but
for now I've duplicated the text.

Then the comment mentions the approvers. Either:
* No one can merge, so we tell all approvers to find someone
  who can.
* One person can merge, ask them only.
* Many can merge, ask all who can.

This leaves the PR approved, without the "no-commit-access" label,
and a notification on its way to the approvers.

There is going to be some delay while the author updates things,
so we could add another round of automation here but for now
it's on the approvers to wait for that.

(you might consider having the author add a label, but of course,
they liklely won't have permissions to do so)

TODO: ...