[GitHub] Add workflows to manage merging of PRs from authors without commit access
These changes aim to make contributors and reviewers aware when a PR
should be merged on the contributor's behalf.

It happens in two parts:
* Newly opened PRs will be labelled with "no-commit-access" if the
  author does not have commit access.
* A new workflow will periodically check all open PRs with this
  label to see if they have approvals, and all checks have finished
  (not passed, just finished, some failures can be explained).
* If they do, it will remove the label and add a comment:
  * Instructing the author to make it merge ready, if needed.
  * Instructing approvers to merge it themselves, or to find
    someone who can.

**Note:** This process could be simplified if we were able to write to the
repository in response to the event generated when an approval is given.
Due to security restrictions in GitHub, we cannot do this, see:
https://github.com/orgs/community/discussions/26651
https://github.com/orgs/community/discussions/55940
Instead, we run the new workflow periodically.

Checking Check status is done using PyGitHub's version of:
https://docs.github.com/en/rest/checks/runs?apiVersion=2022-11-28#list-check-runs-for-a-git-reference

There are some potential corner cases here, but the idea is that
this is enough to make both author and reviewers aware that
merge on behalf is an option. From there, I hope that they can
communicate directly on the PR.

If this does not happen in practice, we can revisit this and
add more automation where it makes sense.