[lldb] Restore register state if PrepareTrivialCall fails
Fixes #124269

PrepareTrivalCall always had the possibility of failing, but
given that it only wrote to general purpose registers, if it
did, you had bigger problems.

When it failed, we did not mark the thread plan valid and
when it was torn down we didn't try to restore the register
state. This meant that if you tried to continue, the program
unlikely to work.

When I added GCS, I needed to handle the situation where the
GCS pointer points to unmapped memory and we fail to write
the extra entry we need. So I added code to restore the gcspr_el0
register specifically if this happened, and ordered the operations
such that we tried this first.

In this change I've made the teardown of an invalid thread plan
restore the register state if one was saved. It may be there isn't
one if ConstructorSetup fails, but this is ok because that function
does not modify anything.

Now that we're doing that, I don't need the GCS specific code
anymore, and all thread plans are protected from this in the rare
event something does fail.

Testing is done by the existing GCS test case that points
the gcspr into unmapped memory which causes PrepareTrivialCall
to fail.

I tried adding a simulated test using a mock gdb server. This
was not possible because they all use DynamicLoaderStatic which
disables all JIT features.