[llvm][release] Reveal download links based on uploaded assets
During the 21.x releases, I've had to hand edit the release page to
reveal the download links, due to the way I added the uncomment download
links step.

It's at the end of release-tasks.yml. Which is good becuase it runs
only once per release (atomic), but bad because it will only run if every
release build on every platform succeeded. For 21.x, MacOS has been
failing always.

To fix this, I wanted to accumulate the succesful builds and only
reval those links.

First attempt was to get the build status from the matrix job that
runs the builds. However GitHub only produces a single status for
a matrix job.

So instead of looking for that, we're going to look directly at
the uploaded release assets. If the build for a given platform
succeeded, it will have uploaded assets.

I have modified the release message content so that each link line
has the form:
```
<!-- DOWNLOAD_LINK * <platform> <links in normal Markdown form> -->
```

uncomment_download_links goes line by line through the message
to find these lines. If all the files a line links to are found in
the uploaded assets, it is edited to be visible.

For instance for lines that link to an installer, an archive, and
signatures, everything must be in the assets.

This means:
* There is one place to find all the links
* There is one place to add new links
* The full links are always there in the release messageif we
  want to hand edit them for some reason.

It is possible that no assets exist and no links are revealed.
To handle this case, the text following the links has been made
more general so it makes sense whether it follows links or not.

This even works for builds that are not automatically uploaded,
though in normal use, those lines will never be revealed because
this command will run before the assets get uploaded.

This command could be run by the release manager or other automation
at other times (perhaps just before making the release immutable, if we
do enable that GitHub feature in future).